requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Django==3.2.18
djangorestframework==3.14.0
Fiona==1.9.4.post1
GDAL==3.4.1
geopandas==0.14.0
install==1.3.5
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
